agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-google-genai>=4.2.2",
    "onnxruntime<1.20.0",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.46.0",
]
